refactor : 메서드 작은 단위로 분할
